+++ src/Resources/ProcessStatus/Link.php
@@ -1,0 +1,11 @@
+<?php
+namespace Budgetlens\BolRetailerApi\Resources\ProcessStatus;
+
+use Budgetlens\BolRetailerApi\Resources\BaseResource;
+
+class Link extends BaseResource
+{
+    public $rel;
+    public $href;
+    public $method;
+}