+++ src/Resources/Replenishment/Warehouse.php
@@ -1,0 +1,14 @@
+<?php
+namespace Budgetlens\BolRetailerApi\Resources\Replenishment;
+
+use Budgetlens\BolRetailerApi\Resources\BaseResource;
+
+class Warehouse extends BaseResource
+{
+    public $streetName;
+    public $houseNumber;
+    public $zipCode;
+    public $city;
+    public $countryCode;
+    public $attentionOf;
+}